Module cppconnect6::cppbind::ffi_test

Test implementation for Rust-C++ FFI.

Test method is implemented in Rust, and it could be called on C++. All functions are not mangled with extern "C" and those are declared in connect6.hpp Reference test_cppbind.

Functions

test_cpp_policy

test_echo_cppeval
Get boards and callback then return the evaluation from CppEval

test_echo_raw_path
Get path info from C++ and return repackaged one.

test_echo_raw_play_result
Get play result from C++ and return repackaged one.

test_echo_raw_vec
Get int array from C++ and return repackaged one.

test_new_raw_path
Return RawPath::new();

test_with_raw_path
Generate sample agent::Path and return RawPath::with_path.

test_with_raw_play_result
Generate sample agent::PlayResult and return RawPlayResult::with_result.

test_with_raw_vec
Generate sample Vec<i32> and return RawVec::with_vec.